- Added a script for dataset 1, which predicts the segmentation of all spheroids, loads the colocalization channel and colocalizes the signal in the colocalization channel, corresponding to the actual label in the segmentation and saves the results and some statistics. - bugfix in the prediction scripts (fixed an uint overflow)

+++ 05-validation/01-ds1-predict_density_maps.py
@@ -77,7 +77,7 @@ for directory in os.listdir(path_to_data):
                                 # Load the ground-truth
                                 centroids_file = os.path.join(res_dir, centroids_filename)
                                 centroids, centroids_header = nrrd.read(centroids_file) #XYZ
-                                num_of_cells_ground_truth = np.sum(centroids)
+                                num_of_cells_ground_truth = np.sum(centroids).astype(np.float)
     
                                 # Predict the density-map
                                 spheroid_new, density_map, num_of_cells_predicted = cnn.predict_density_map(path_to_spheroid=spheroid_file, patch_sizes=patch_sizes,
